+++ app/app/not-found.tsx
@@ -3,5 +3,5 @@
 import NotFoundView from "@/components/ui/not-found-view";
 
 export default function NotFound() {
-  return <NotFoundView isApp={false} />;
+  return <NotFoundView isApp={true} />;
 }